refactor: split uploading
